+++ src/Botc/Logger/index.ts
@@ -1,0 +1,7 @@
+import { Logger } from './Logger.js';
+import { LogLevel } from './Logger.types.js';
+
+export {
+  Logger,
+  LogLevel,
+};