+++ src/app.ts
@@ -1,11 +1,13 @@
 import { Botc } from './Botc/index.js';
+import { Logger } from './Botc/Logger/Logger.js';
 
 try {
   new Botc();
 }
 catch (error) {
   if (error instanceof Error) {
-    console.error(error.message);
-    console.error(error.stack);
+    const logger = new Logger();
+    logger.log(error.message, 'ERROR');
+    logger.log(error.stack || 'Stack unavailable', 'ERROR');
   }
 }